Remove IN_MODIFY, IN_CREATE
There is a badarg exception in
binary_languages_utils:consult(blah, "path/to/blah.bert"). It is possible
that a race exist between blah.bert being fully written to and filewatch sending
a message to an rtb_lib_indexer to reload the file. This is an attempt to debug
this.